requirements.txt
langchain
langchain-community
langchain-google-genai
google-search-results
streamlit
python-dotenv
